refractoring, separated the adding and deleting properties of AddImageButton into Interface
